Update dependencies and add additiona file structure
