+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.6
+// swift-tools-version:5.7
 
 //
 // This source file is part of the ResearchKitOnFHIR open source project
@@ -21,7 +21,7 @@ let package = Package(
         .library(name: "FHIRQuestionnaires", targets: ["FHIRQuestionnaires"])
     ],
     dependencies: [
-        .package(url: "https://github.com/PSchmiedmayer/ResearchKit.git", from: "2.1.2"),
+        .package(url: "https://github.com/CardinalKit/ResearchKit.git", from: "2.1.1"),
         .package(url: "https://github.com/apple/FHIRModels.git", .upToNextMajor(from: "0.4.0"))
     ],
     targets: [
